Refactor AutoBalancer into composable parts
